+++ TheGame/__init__.py
@@ -1,3 +1,3 @@
 from TheGame import Models
 from TheGame.World.Environment import Environment
-from TheGame import utils
+from TheGame import Results